pyproject.toml
[project]
name = "fastapi-task-manager"
version = "0.1.0"
description = ""
authors = [{ name = "Powermacintosh", email = "ak.powermacintosh@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "loki-logger-handler (>=1.1.2,<2.0.0)",
]

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=api_v1 --cov-report=term-missing"
python_files = "test_*.py"
env_files = [".test.env"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
pytest-cov = "^6.2.1"
pytest-dotenv = "^0.5.2"
